Refactor BackgroundGlow and FancyButton components; update imports and styles
