+++ TODO.txt
@@ -1,6 +1,3 @@
-Fancy button
-Balanced text
-
 Row Level Security
 app.cogniba.com
 Landing page
