+++ src/components/ui/BackgroundGlow/BackgroundGlow.module.css
@@ -1,0 +1,28 @@
+@property --angle {
+  syntax: "<angle>";
+  initial-value: 0deg;
+  inherits: false;
+}
+
+.backgroundGlow::after {
+  content: "";
+  position: absolute;
+  height: 100%;
+  width: 100%;
+  top: 50%;
+  left: 50%;
+  translate: -50% -50%;
+  z-index: -1;
+  background-image: conic-gradient(from var(--angle), var(--glow-colors));
+  animation: var(--animation-duration) spin linear infinite;
+  filter: blur(var(--glow-size));
+}
+
+@keyframes spin {
+  from {
+    --angle: 0deg;
+  }
+  to {
+    --angle: 360deg;
+  }
+}
